Abstracted UsersController to IUsersManager
