+++ Registry.Web/Services/Ports/IUsersManager.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
+using Registry.Web.Models;
+using Registry.Web.Models.DTO;
+
+namespace Registry.Web.Services.Ports
+{
+    public interface IUsersManager
+    {
+        Task<AuthenticateResponse> Authenticate(AuthenticateRequest model);
+        Task<IEnumerable<UserDto>> GetAll();
+    }
+}
