+++ Registry.Web/Startup.cs
@@ -127,6 +127,7 @@ namespace Registry.Web
             services.AddScoped<IUtils, WebUtils>();
             services.AddScoped<IAuthManager, AuthManager>();
 
+            services.AddScoped<IUsersManager, UsersManager>();
             services.AddScoped<IOrganizationsManager, OrganizationsManager>();
             services.AddScoped<IDatasetsManager, DatasetsManager>();
             services.AddScoped<IObjectsManager, ObjectsManager>();
